lint(vx-glyph)
