+++ packages/vx-glyph/src/glyphs/GlyphStar.jsx
@@ -1,25 +1,25 @@
 import React from 'react';
 import cx from 'classnames';
 import PropTypes from 'prop-types';
-import { symbol, symbolCircle } from 'd3-shape';
+import { symbol, symbolStar } from 'd3-shape';
 import Glyph from './Glyph';
 
-GlyphCircle.propTypes = {
+GlyphStar.propTypes = {
   children: PropTypes.func,
   className: PropTypes.string,
   top: PropTypes.number,
   left: PropTypes.number,
-  size: PropTypes.oneOfType([PropTypes.number, PropTypes.func])
+  size: PropTypes.oneOfType([PropTypes.number, PropTypes.func]),
 };
 
-export default function GlyphCircle({ children, className, top, left, size, ...restProps }) {
+export default function GlyphStar({ children, className, top, left, size, ...restProps }) {
   const path = symbol();
-  path.type(symbolCircle);
+  path.type(symbolStar);
   if (size) path.size(size);
   if (children) return children({ path });
   return (
     <Glyph top={top} left={left}>
-      <path className={cx('vx-glyph-circle', className)} d={path()} {...restProps} />
+      <path className={cx('vx-glyph-star', className)} d={path()} {...restProps} />
     </Glyph>
   );
 }
