+++ packages/vx-glyph/src/glyphs/GlyphTriangle.jsx
@@ -1,25 +1,25 @@
 import React from 'react';
 import cx from 'classnames';
 import PropTypes from 'prop-types';
-import { symbol, symbolSquare } from 'd3-shape';
+import { symbol, symbolTriangle } from 'd3-shape';
 import Glyph from './Glyph';
 
-GlyphSquare.propTypes = {
+GlyphTriangle.propTypes = {
   children: PropTypes.func,
   className: PropTypes.string,
   top: PropTypes.number,
   left: PropTypes.number,
-  size: PropTypes.oneOfType([PropTypes.number, PropTypes.func])
+  size: PropTypes.oneOfType([PropTypes.number, PropTypes.func]),
 };
 
-export default function GlyphSquare({ children, className, top, left, size, ...restProps }) {
+export default function GlyphTriangle({ children, className, top, left, size, ...restProps }) {
   const path = symbol();
-  path.type(symbolSquare);
+  path.type(symbolTriangle);
   if (size) path.size(size);
   if (children) return children({ path });
   return (
     <Glyph top={top} left={left}>
-      <path className={cx('vx-glyph-square', className)} d={path()} {...restProps} />
+      <path className={cx('vx-glyph-triangle', className)} d={path()} {...restProps} />
     </Glyph>
   );
 }
