+++ packages/vx-glyph/test/Circle.test.jsx
@@ -1,29 +1,32 @@
-import { GlyphSquare } from '../src';
+import React from 'react';
+import { shallow } from 'enzyme';
 
-describe('<GlyphSquare />', () => {
+import { GlyphCircle } from '../src';
+
+describe('<GlyphCircle />', () => {
   test('it should be defined', () => {
-    expect(GlyphSquare).toBeDefined();
+    expect(GlyphCircle).toBeDefined();
   });
 
   test('it should be wrapped in a <Glyph />', () => {
-    const wrapper = shallow(<GlyphSquare />);
+    const wrapper = shallow(<GlyphCircle />);
     expect(wrapper.dive().prop('className')).toBe('vx-glyph');
   });
 
   test('it should add className to <path />', () => {
-    const wrapper = shallow(<GlyphSquare className="test" />);
-    expect(wrapper.find('.test').length).toBe(1);
+    const wrapper = shallow(<GlyphCircle className="test" />);
+    expect(wrapper.find('.test')).toHaveLength(1);
   });
 
   test('it should take a children as function prop', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphSquare>{fn}</GlyphSquare>);
+    shallow(<GlyphCircle>{fn}</GlyphCircle>);
     expect(fn).toHaveBeenCalled();
   });
 
   test('it should call children function with { path }', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphSquare>{fn}</GlyphSquare>);
+    shallow(<GlyphCircle>{fn}</GlyphCircle>);
     const args = fn.mock.calls[0][0];
     const keys = Object.keys(args);
     expect(keys.includes('path')).toEqual(true);
@@ -31,7 +34,7 @@ describe('<GlyphSquare />', () => {
 
   test('it should take a size prop as a number', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphSquare size={42}>{fn}</GlyphSquare>);
+    shallow(<GlyphCircle size={42}>{fn}</GlyphCircle>);
     const args = fn.mock.calls[0][0];
     expect(args.path.size()()).toBe(42);
   });
@@ -39,7 +42,7 @@ describe('<GlyphSquare />', () => {
   test('it should take a size prop as a function', () => {
     const fn = jest.fn();
     const sizeFn = () => 42;
-    const wrapper = shallow(<GlyphSquare size={sizeFn}>{fn}</GlyphSquare>);
+    shallow(<GlyphCircle size={sizeFn}>{fn}</GlyphCircle>);
     const args = fn.mock.calls[0][0];
     expect(args.path.size()()).toBe(42);
   });
