+++ packages/vx-glyph/test/Cross.test.jsx
@@ -1,29 +1,32 @@
-import { GlyphCircle } from '../src';
+import React from 'react';
+import { shallow } from 'enzyme';
 
-describe('<GlyphCircle />', () => {
+import { GlyphCross } from '../src';
+
+describe('<GlyphCross />', () => {
   test('it should be defined', () => {
-    expect(GlyphCircle).toBeDefined();
+    expect(GlyphCross).toBeDefined();
   });
 
   test('it should be wrapped in a <Glyph />', () => {
-    const wrapper = shallow(<GlyphCircle />);
+    const wrapper = shallow(<GlyphCross />);
     expect(wrapper.dive().prop('className')).toBe('vx-glyph');
   });
 
   test('it should add className to <path />', () => {
-    const wrapper = shallow(<GlyphCircle className="test" />);
-    expect(wrapper.find('.test').length).toBe(1);
+    const wrapper = shallow(<GlyphCross className="test" />);
+    expect(wrapper.find('.test')).toHaveLength(1);
   });
 
   test('it should take a children as function prop', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphCircle>{fn}</GlyphCircle>);
+    shallow(<GlyphCross>{fn}</GlyphCross>);
     expect(fn).toHaveBeenCalled();
   });
 
   test('it should call children function with { path }', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphCircle>{fn}</GlyphCircle>);
+    shallow(<GlyphCross>{fn}</GlyphCross>);
     const args = fn.mock.calls[0][0];
     const keys = Object.keys(args);
     expect(keys.includes('path')).toEqual(true);
@@ -31,7 +34,7 @@ describe('<GlyphCircle />', () => {
 
   test('it should take a size prop as a number', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphCircle size={42}>{fn}</GlyphCircle>);
+    shallow(<GlyphCross size={42}>{fn}</GlyphCross>);
     const args = fn.mock.calls[0][0];
     expect(args.path.size()()).toBe(42);
   });
@@ -39,7 +42,7 @@ describe('<GlyphCircle />', () => {
   test('it should take a size prop as a function', () => {
     const fn = jest.fn();
     const sizeFn = () => 42;
-    const wrapper = shallow(<GlyphCircle size={sizeFn}>{fn}</GlyphCircle>);
+    shallow(<GlyphCross size={sizeFn}>{fn}</GlyphCross>);
     const args = fn.mock.calls[0][0];
     expect(args.path.size()()).toBe(42);
   });
