+++ packages/vx-glyph/test/Dot.test.jsx
