+++ packages/vx-glyph/test/Square.test.jsx
@@ -1,29 +1,32 @@
-import { GlyphCross } from '../src';
+import React from 'react';
+import { shallow } from 'enzyme';
 
-describe('<GlyphCross />', () => {
+import { GlyphSquare } from '../src';
+
+describe('<GlyphSquare />', () => {
   test('it should be defined', () => {
-    expect(GlyphCross).toBeDefined();
+    expect(GlyphSquare).toBeDefined();
   });
 
   test('it should be wrapped in a <Glyph />', () => {
-    const wrapper = shallow(<GlyphCross />);
+    const wrapper = shallow(<GlyphSquare />);
     expect(wrapper.dive().prop('className')).toBe('vx-glyph');
   });
 
   test('it should add className to <path />', () => {
-    const wrapper = shallow(<GlyphCross className="test" />);
-    expect(wrapper.find('.test').length).toBe(1);
+    const wrapper = shallow(<GlyphSquare className="test" />);
+    expect(wrapper.find('.test')).toHaveLength(1);
   });
 
   test('it should take a children as function prop', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphCross>{fn}</GlyphCross>);
+    shallow(<GlyphSquare>{fn}</GlyphSquare>);
     expect(fn).toHaveBeenCalled();
   });
 
   test('it should call children function with { path }', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphCross>{fn}</GlyphCross>);
+    shallow(<GlyphSquare>{fn}</GlyphSquare>);
     const args = fn.mock.calls[0][0];
     const keys = Object.keys(args);
     expect(keys.includes('path')).toEqual(true);
@@ -31,7 +34,7 @@ describe('<GlyphCross />', () => {
 
   test('it should take a size prop as a number', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphCross size={42}>{fn}</GlyphCross>);
+    shallow(<GlyphSquare size={42}>{fn}</GlyphSquare>);
     const args = fn.mock.calls[0][0];
     expect(args.path.size()()).toBe(42);
   });
@@ -39,7 +42,7 @@ describe('<GlyphCross />', () => {
   test('it should take a size prop as a function', () => {
     const fn = jest.fn();
     const sizeFn = () => 42;
-    const wrapper = shallow(<GlyphCross size={sizeFn}>{fn}</GlyphCross>);
+    shallow(<GlyphSquare size={sizeFn}>{fn}</GlyphSquare>);
     const args = fn.mock.calls[0][0];
     expect(args.path.size()()).toBe(42);
   });
