+++ packages/vx-glyph/test/Star.test.jsx
@@ -1,3 +1,6 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+
 import { GlyphStar } from '../src';
 
 describe('<GlyphStar />', () => {
@@ -12,18 +15,18 @@ describe('<GlyphStar />', () => {
 
   test('it should add className to <path />', () => {
     const wrapper = shallow(<GlyphStar className="test" />);
-    expect(wrapper.find('.test').length).toBe(1);
+    expect(wrapper.find('.test')).toHaveLength(1);
   });
 
   test('it should take a children as function prop', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphStar>{fn}</GlyphStar>);
+    shallow(<GlyphStar>{fn}</GlyphStar>);
     expect(fn).toHaveBeenCalled();
   });
 
   test('it should call children function with { path }', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphStar>{fn}</GlyphStar>);
+    shallow(<GlyphStar>{fn}</GlyphStar>);
     const args = fn.mock.calls[0][0];
     const keys = Object.keys(args);
     expect(keys.includes('path')).toEqual(true);
@@ -31,7 +34,7 @@ describe('<GlyphStar />', () => {
 
   test('it should take a size prop as a number', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphStar size={42}>{fn}</GlyphStar>);
+    shallow(<GlyphStar size={42}>{fn}</GlyphStar>);
     const args = fn.mock.calls[0][0];
     expect(args.path.size()()).toBe(42);
   });
@@ -39,7 +42,7 @@ describe('<GlyphStar />', () => {
   test('it should take a size prop as a function', () => {
     const fn = jest.fn();
     const sizeFn = () => 42;
-    const wrapper = shallow(<GlyphStar size={sizeFn}>{fn}</GlyphStar>);
+    shallow(<GlyphStar size={sizeFn}>{fn}</GlyphStar>);
     const args = fn.mock.calls[0][0];
     expect(args.path.size()()).toBe(42);
   });
