+++ packages/vx-glyph/test/Wye.test.jsx
@@ -1,29 +1,32 @@
-import { GlyphDiamond } from '../src';
+import React from 'react';
+import { shallow } from 'enzyme';
 
-describe('<GlyphDiamond />', () => {
+import { GlyphWye } from '../src';
+
+describe('<GlyphWye />', () => {
   test('it should be defined', () => {
-    expect(GlyphDiamond).toBeDefined();
+    expect(GlyphWye).toBeDefined();
   });
 
   test('it should be wrapped in a <Glyph />', () => {
-    const wrapper = shallow(<GlyphDiamond />);
+    const wrapper = shallow(<GlyphWye />);
     expect(wrapper.dive().prop('className')).toBe('vx-glyph');
   });
 
   test('it should add className to <path />', () => {
-    const wrapper = shallow(<GlyphDiamond className="test" />);
-    expect(wrapper.find('.test').length).toBe(1);
+    const wrapper = shallow(<GlyphWye className="test" />);
+    expect(wrapper.find('.test')).toHaveLength(1);
   });
 
   test('it should take a children as function prop', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphDiamond>{fn}</GlyphDiamond>);
+    shallow(<GlyphWye>{fn}</GlyphWye>);
     expect(fn).toHaveBeenCalled();
   });
 
   test('it should call children function with { path }', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphDiamond>{fn}</GlyphDiamond>);
+    shallow(<GlyphWye>{fn}</GlyphWye>);
     const args = fn.mock.calls[0][0];
     const keys = Object.keys(args);
     expect(keys.includes('path')).toEqual(true);
@@ -31,7 +34,7 @@ describe('<GlyphDiamond />', () => {
 
   test('it should take a size prop as a number', () => {
     const fn = jest.fn();
-    const wrapper = shallow(<GlyphDiamond size={42}>{fn}</GlyphDiamond>);
+    shallow(<GlyphWye size={42}>{fn}</GlyphWye>);
     const args = fn.mock.calls[0][0];
     expect(args.path.size()()).toBe(42);
   });
@@ -39,7 +42,7 @@ describe('<GlyphDiamond />', () => {
   test('it should take a size prop as a function', () => {
     const fn = jest.fn();
     const sizeFn = () => 42;
-    const wrapper = shallow(<GlyphDiamond size={sizeFn}>{fn}</GlyphDiamond>);
+    shallow(<GlyphWye size={sizeFn}>{fn}</GlyphWye>);
     const args = fn.mock.calls[0][0];
     expect(args.path.size()()).toBe(42);
   });
